2021-01-03 - Watch-DSClientActivity Created new Cmdlet Watch-DSClientActivity providing a Progress Indicator for a specified running activity Created a new Base class for DSClientAcitvityLog
